generator fixes
